+++ CHANGES.md
@@ -5,7 +5,7 @@
 - Fixed broken point dragging interaction for user drawing in 3D mode.
 - Fixed rectangle drawing in 2D mode.
 - Added EPSG:7855 to `Proj4Definitions`.
-- Support time interval query via model dimensions for esri-mapServer type.
+- Add time window duration trait. The esri-mapServer type supports time window duration query via model dimensions.
 - [The next improvement]
 
 #### 8.3.2 - 2023-08-11